Improved code organization
